fix: R2 & 3PO causing tons of assertion errors by existing
The original way I made their space UniqueUnit was by copying the Gallofree Transport and setting the Millennium Falcon as a base.  Now, with the Falcon as a base and filled with Gallofree values, I can achieve the same effect without causing 1,000 assertions b/c I'm overwriting an ability of a variant (which is a no-no).